requirements.txt
tensorflow
torchmetrics
mlflow
startlette
pytorch
torchvision
ray[serve]
